requirements.txt
# Зависимости веб-приложения MedTech AI (Flask)
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.3
flask-wtf>=1.2.0
email-validator>=2.1.0
werkzeug>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
